Added L2 to L1 cache data flow controlled via a flag so it can be set on demand.
